:boom: :burn: Typescript integrated into debugger, cli bin refactor, closes #691
