+++ .vscode/launch.json
@@ -1,23 +1,20 @@
 {
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "node",
-            "request": "launch",
-            "name": "Launch Program",
-            "program": "${workspaceFolder}/bin/mwoffliner.script.js",
-            "args": [
-                "--mwUrl=https://bm.wikipedia.org",
-                "--adminEmail=admin@kiwix.com",
-                "--verbose",
-                "--format=nozim",
-                // "--articleList=./articleList",
-                "--redis=redis://127.0.0.1:6379"
-                // "--customZimFavicon=test.png"
-            ]
-        }
-    ]
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "name": "Debug",
+      "type": "node",
+      "request": "launch",
+      "cwd": "${workspaceRoot}",
+      "runtimeExecutable": "npm",
+      "runtimeArgs": [
+        "run",
+        "debug"
+      ],
+      "port": 9229
+    }
+  ]
 }
