+++ build.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 echo "Building at [$(date)]"
 npx tsc
+npx replace "':'; //#" "':' //#" ./lib/cli.js 
 echo "Build Complete at [$(date)]"
